Move datasource.save to sdk
